+++ doc/sample_state_machine_specification.csv
@@ -4,7 +4,7 @@ test_state,0,"booleanAnd(booleanNot(booleanOr(keyword(""dont""),keyword(""don't"
 further_details_access_question,0,"or(and(or(keyword(""problem.*""),keyword(""issue.*""),keyword(""trouble.*"")),keyword(""account"")),search(""further_details_access_question""))","[""cannot access account"", ""problem access account""]","Hello and welcome to our customer service chat. Please note that while I am not a human operator, I will do my very best to assist You today. How may I help you?",show_buttons,"{""Forgot Password"": ""forgot_password"", ""Account locked"": ""account_locked"", ""Specify your problem"": ""specify_problem"", ""I want to call an operator"": ""call_operator"", ""None of the above"": ""start""}","{""verification"": ""did you mean you can't access to your account?""}",eval(show_buttons),"dont_understand"
 forgot_password,0,"and(or(keyword(""reset""),keyword(""forgot"")),keyword(""password""))",,We can reset your password by sending you a message to your registered e-mail address. Please type your email address:,input_form,"{""email"": ""email""}","{""verification"": ""did you mean you forgot the password?""}","send_password_generation_link","dont_understand"
 send_password_generation_link,0,,,Thank you. An e-mail will be sent to this address: %email% with your account details and the necessary steps for you to reset your password.,send_password_generation_link,"{ ""template"": ""Hi,\nSomeone requested a new password for your account. You can set a new password here: %link%\nIf you did not request this, just ignore this message."", ""email"": ""%email%"", ""subject"": ""New password"" }",,any_further,call_operator
-account_locked,0,"booleanand(keyword(""locked""), keyword(""account""))",,"If your account is locked you need to call our customer service at 800 000 000",,,,any_further,
+account_locked,0,"booleanand(keyword(""locked""), keyword(""account""), )",,"If your account is locked you need to call our customer service at 800 000 000",,,,any_further,
 goodbye_question,0,,,"Would you mind rating this conversation so my developers know if they're doing a good job?",show_buttons,"{ ""Excellent"": ""excellent_feedback"", ""Terrible"": ""terrible_feedback"" }",,eval(show_buttons),dont_understand
 excellent_feedback,0,,,"Great!, See you soon!",,,,,
 terrible_feedback,0,"booleanor(keyword(""idiot""), keyword(""fuck.*""), keyword(""screw""), keyword(""damn.*""), keyword(""asshole""))",,"Sorry to hear that. I'll put you in touch with an operator",,,,call_operator,
