+++ src/main/scala/com/getjenny/starchat/services/DecisionTableService.scala
@@ -38,213 +38,6 @@ import org.elasticsearch.action.admin.indices.refresh.RefreshResponse
 class DecisionTableService(implicit val executionContext: ExecutionContext) {
   val elastic_client = DecisionTableElasticClient
   val log: LoggingAdapter = Logging(SCActorSystem.system, this.getClass.getCanonicalName)
-  val termService = new TermService
-
-  case class AnalyzerItem(declaration: String,
-                        build: Boolean,
-                        analyzer: StarchatAnalyzer,
-                        queries: List[TextTerms])
-
-  var analyzer_map : Map[String, AnalyzerItem] = Map.empty[String, AnalyzerItem]
-
-  def getAnalyzers: Map[String, AnalyzerItem] = {
-    val client: TransportClient = elastic_client.get_client()
-    val qb : QueryBuilder = QueryBuilders.matchAllQuery()
-
-    val refresh_index = elastic_client.refresh_index()
-    if(refresh_index.failed_shards_n > 0) {
-      throw new Exception("DecisionTable : index refresh failed: (" + elastic_client.index_name + ")")
-    }
-
-    val scroll_resp : SearchResponse = client.prepareSearch(elastic_client.index_name)
-      .setTypes(elastic_client.type_name)
-      .setQuery(qb)
-      .setFetchSource(Array("state", "analyzer", "queries"), Array.empty[String])
-      .setScroll(new TimeValue(60000))
-      .setSize(1000).get()
-
-    //get a map of stateId -> AnalyzerItem (only if there is smt in the field "analyzer")
-    val results : Map[String, AnalyzerItem] = scroll_resp.getHits.getHits.toList.map({ e =>
-      val item: SearchHit = e
-      val state : String = item.getId
-      val source : Map[String, Any] = item.getSource.asScala.toMap
-      val declaration : String = source.get("analyzer") match {
-        case Some(t) => t.asInstanceOf[String]
-        case None => ""
-      }
-
-      val analyzer : StarchatAnalyzer = if (declaration != "") {
-        new StarchatAnalyzer(declaration)
-      } else {
-        null
-      }
-
-      val build = analyzer != null
-
-      val queries : List[String] = source.get("queries") match {
-        case Some(t) => t.asInstanceOf[java.util.ArrayList[String]].asScala.toList
-        case None => List[String]()
-      }
-
-      val queries_terms: List[TextTerms] = queries.map(q => {
-        val query_terms = termService.textToVectors(q)
-        query_terms
-      }).filter(_.nonEmpty).map(x => x.get)
-
-      val analyzerItem = AnalyzerItem(declaration, build, analyzer, queries_terms)
-      (state, analyzerItem)
-    }).filter(_._2.declaration != "").toMap
-    results
-  }
-
-  def loadAnalyzer : Future[Option[DTAnalyzerLoad]] = Future {
-    analyzer_map = getAnalyzers
-    val dt_analyzer_load = DTAnalyzerLoad(num_of_entries=analyzer_map.size)
-    Option {dt_analyzer_load}
-  }
-
-  def getDTAnalyzerMap : Future[Option[DTAnalyzerMap]] = {
-    val analyzers = Future(Option(DTAnalyzerMap(analyzer_map.map(x => {
-      val dt_analyzer = DTAnalyzerItem(x._2.declaration, x._2.build)
-      (x._1, dt_analyzer)
-    }))))
-    analyzers
-  }
-
-  def initializeAnalyzers(): Unit = {
-    val result: Try[Option[DTAnalyzerLoad]] =
-      Await.ready(loadAnalyzer, 30.seconds).value.get
-    result match {
-      case Success(t) => log.info("analyzers loaded")
-      case Failure(e) => log.error("can't load analyzers: " + e.toString)
-    }
-  }
-
-  initializeAnalyzers() // load analyzer map on startup
-
-  def getNextResponse(request: ResponseRequestIn): Option[ResponseRequestOutOperationResult] = {
-    // calculate and return the ResponseRequestOut
-
-    val user_text: String = if (request.user_input.isDefined) {
-      request.user_input.get.text.getOrElse("")
-    } else {
-      ""
-    }
-
-    val conversation_id: String = request.conversation_id
-
-    val data: Map[String, String] = if (request.values.isDefined)
-      request.values.get.data.getOrElse(Map[String, String]())
-    else
-      Map[String, String]()
-
-    val return_value: String = if (request.values.isDefined)
-      request.values.get.return_value.getOrElse("")
-    else
-      ""
-
-    val return_state: Option[ResponseRequestOutOperationResult] = Option {
-      if (!return_value.isEmpty) {
-        // there is a state in return_value (eg the client asked for a state), no analyzers evaluation
-        val state: Future[Option[SearchDTDocumentsResults]] = read(List[String](return_value))
-        val res: Option[SearchDTDocumentsResults] = Await.result(state, 30.seconds)
-        if (res.get.total > 0) {
-          val doc: DTDocument = res.get.hits.head.document
-          val state: String = doc.state
-          val max_state_count: Int = doc.max_state_count
-          val analyzer: String = doc.analyzer
-          var bubble: String = doc.bubble
-          var action_input: Map[String, String] = doc.action_input
-          val state_data: Map[String, String] = doc.state_data
-          if (data.nonEmpty) {
-            for ((key, value) <- data) {
-              bubble = bubble.replaceAll("%" + key + "%", value)
-              action_input = doc.action_input map { case (ki, vi) =>
-                val new_value: String = vi.replaceAll("%" + key + "%", value)
-                (ki, new_value)
-              }
-            }
-          }
-
-          val response_data: ResponseRequestOut = ResponseRequestOut(conversation_id = conversation_id,
-            state = state,
-            max_state_count = max_state_count,
-            analyzer = analyzer,
-            bubble = bubble,
-            action = doc.action,
-            data = data,
-            action_input = action_input,
-            state_data = state_data,
-            success_value = doc.success_value,
-            failure_value = doc.failure_value,
-            score = 1.0d)
-
-          val full_response: ResponseRequestOutOperationResult =
-            ResponseRequestOutOperationResult(ReturnMessageData(200, ""), Option {
-              List(response_data)
-            }) // success
-          full_response
-        } else {
-          val full_response: ResponseRequestOutOperationResult =
-            ResponseRequestOutOperationResult(ReturnMessageData(500,
-              "Error during state retrieval"), null) // internal error
-          full_response
-        }
-      } else {
-        // No states in the return values
-        val max_results: Int = request.max_results.getOrElse(2)
-        val threshold: Double = request.threshold.getOrElse(0.0d)
-        val analyzer_values: Map[String, Double] = analyzer_map.filter(_._2.build == true).map(item => {
-          val evaluation_score = item._2.analyzer.evaluate(user_text)
-          val state_id = item._1
-          (state_id, evaluation_score)
-        }).toList.filter(_._2 > threshold).sortWith(_._2 > _._2).take(max_results).toMap
-
-        if(analyzer_values.nonEmpty) {
-          val items: Future[Option[SearchDTDocumentsResults]] = read(analyzer_values.keys.toList)
-          val res : Option[SearchDTDocumentsResults] = Await.result(items, 30.seconds)
-          val docs = res.get.hits.map(item => {
-            val doc: DTDocument = item.document
-            val state = doc.state
-            val score: Double = analyzer_values(state)
-            val max_state_count: Int = doc.max_state_count
-            val analyzer: String = doc.analyzer
-            var bubble: String = doc.bubble
-            var action_input: Map[String, String] = doc.action_input
-            val state_data: Map[String, String] = doc.state_data
-
-            if (data.nonEmpty) {
-              for ((key, value) <- data) {
-                bubble = bubble.replaceAll("%" + key + "%", value)
-                action_input = doc.action_input map { case (ki, vi) =>
-                  val new_value: String = vi.replaceAll("%" + key + "%", value)
-                  (ki, new_value)
-                }
-              }
-            }
-
-            val response_item: ResponseRequestOut = ResponseRequestOut(conversation_id = conversation_id,
-              state = state,
-              max_state_count = max_state_count,
-              analyzer = analyzer,
-              bubble = bubble,
-              action = doc.action,
-              data = data,
-              action_input = action_input,
-              state_data = state_data,
-              success_value = doc.success_value,
-              failure_value = doc.failure_value,
-              score = score)
-            response_item
-          }).sortWith(_.score > _.score)
-          ResponseRequestOutOperationResult(ReturnMessageData(200, ""), Option{docs}) // success
-        } else {
-          ResponseRequestOutOperationResult(ReturnMessageData(204, ""), null)  // no data
-        }
-      }
-    }
-    return_state
-  }
 
   def search(documentSearch: DTDocumentSearch): Future[Option[SearchDTDocumentsResults]] = {
     val client: TransportClient = elastic_client.get_client()
